# Conflicts:
#	Test/D3D12/D3D12Renderer.cpp

+++ Test/D3D12/D3D12Renderer.hpp
@@ -47,7 +47,6 @@ public:
 
 	ID3D12Device4* GetDevice() const;
 	ID3D12RootSignature* GetRootSignature() const;
-	//ID3D12GraphicsCommandList3* GetCommandList() const;
 	D3D12TextureLoader* GetTextureLoader() const;
 
 private:
@@ -93,7 +92,7 @@ private:
 	void SetUpRenderInstructions();
 	void ResetCommandListAndAllocator(int backbufferIndex, int index);
 	void MapMatrixData(int backBufferIndex);
-	void RecordRenderInstructions(D3D12Window* w, D3D12Camera* c, int commandListIndex, int backBufferIndex, int firstInstructionIndex, int numInstructions);
+	void RecordRenderInstructions(D3D12Window* w, D3D12Camera* c, int commandListIndex, int backBufferIndex, size_t firstInstructionIndex, size_t numInstructions);
 
 	void RecordCommands(int threadIndex);
 	void SetThreadWork(int threadIndex, D3D12Window* w, D3D12Camera* c, int backBufferIndex, int firstInstructionIndex, int numInstructions);
